2780 infants on tablet flights
We've switched up the infant seating preference messaging to the user. Instead of
hiding it away in the filters fragment (which conceptually didn't make much sense,
anyway), we now pop a dialog informing the user of our default setting and offer
the ability to change from that default, being clear all the way!