Add save/overwrite dialogs to travelers when logged in.
This is not totally correct as it sits right now. We have no way to explicitly create a new traveler, so if we are logged
in and we change the name of one of our travelers, that will be updated (instead of creating new traveler with a new name).

This whole flow is not great, we arent waiting for the save to complete, we arent updating the autocomplete list, things are
still not great, but atleast the feature is in now and we can fix it up.