SimpleCallbackDialogFragment callback id refactor
Now that SimpleCallbackDialogFragment no longer uses a simple listener interface
we need to make sure the callback id is globally unique. Because the Otto event
bus is more or less global, some strange bugs could come up in the middle of
development if you happen to conflict on the id. By keeping all of the int
definitions in one place, we can mantain our sanity.