Unify a bunch of copy paste code.
At checkout time we need to migrate some of our data from our traveler to BillingInfo, we had some
pretty similar code doing this on both sides of the phone flow, and then it seemed to have been copied
willy nilly to the tablet side. Now we unify the implementation and in the case of tablet we copy data
at a much more appropriate time.