Update a layout XML for tablet flight list
After Doug's recent refactor to the list rows for hotels, flights, and
trip bucket, this XML for the selected flight leg in the header (with
the checkmark) needed to get updated.

This commit copies the XML code from the standard row XML and replaces
the price TextView with an ImageView for the checkmark.