Update another layout XML for tablet flight list
This is similar to the previous commit. There was another layout XML that was similar to the
standard flight row, this one being the XML for the "add to trip" animation. I think there
will be some more cleanup necessary for this down the line (it is using the special roundtrip
view of the lines and circles which is no longer in the designs to the best of my knowledge),
and we also want to get rid of the white border for this animation. At the very least, we have
removed an unnecessary layout for now.