Update children age spinner styling
This commit may not be the best commit. This changes an XML file that is being used
in a few other spinners throughout the app, although, I think the Roboto styling
should be fine.

According to the new mocks, we don't bold the age in the spinner. Rather than changing the string
resource definitions, I just strip out the HTML formatting.