2830 - Fix selecting wrong traveler bug.
The issue: When we select an item from the dropdown, it populates the dropdown field (via Filterable.convertResultToString)
and since we are filtering our results based on the text in the dropdown field, selecting an item would be changing the result
set so when we called adapter.getItem(position) the filtered data would often be different than what the user saw, so selecting
item #2 would result in selecting item #2 after filtering on the name of the selection.

To solve the issue we use stable ids (traveler tuids) to select the correct traveler consistently.