Introduce GuestPicker - a custom View for picking guests!
The design for picking adults, children, and ages on tablet is being
ported over to phone so I felt the need to unify all this logic into
one thing. The architecture is largely inspired by CalendarPicker,
although, there's no custom drawing with GuestPicker.

A lot of the guest picking logic was lifted from ResultsGuestPicker as
it had already been written. Soon to be ported over to tablet and phone
implementations...