ExpediaImageManager crash fix
I came across this crash during dev and it has shown up in HockeyApp crash logs. It
looks like the HTTP request is bombing out with IOException for some reason. In this
case we want to just return a null bitmap instead of crashing. My guess is the servers
are being flaky.