Improve loading list animations on Tablet Results.
Prior to this commit we were occasionally laying out during the animation, this was caused by the text visibility change
which would happen if the number of loading animation listeners changed during the animation, which happened often.

Now we just set our text visibility via alpha to avoid layout, and we wait to re-register for loading bar updates
so that if the fragment gets detached shortly after our expanding animation we never re-register for loading bar updates.