2358 - Download associated traveler details.
So the login/update user api call does not include detailed info about associated travelers, only names
and tuids, we have to do additional network calls to get the additional info. In the past we waited until
the user selected a traveler to download those additional details, however we want to display some of
that data before the user clicks on a traveler, so we need to do the DL up front. We will still show a
spinner and download the traveler on click time if it has not yet finished downloading.