Fighting the OOM fight
HockeyApp crash logs show OOM crashes when saving the FlightSearch
response to disk after it has been parsed. This makes sense as the
this object tends to be pretty large and the way we persist necessitates
converting this response to JSON, all in memory, and then writing
to disk. Essentially, 2x response size is in memory when saving to disk
and boom.

We just catch the OOM and report failure. Disk persistence is not mission critical.