Hack to fix gap at bottom of screen when scrolling up
Due to the expanding Mobile Exclusive Deals on hotel fragments, scrolling
the list up from the main search results screen sometimes resulted in a
transparent gap under the second item (up to maybe 20dp) for a while. Filling
the background of the bottom half of the list with a white background
makes that less obvious.