ExpediaImageManager crash fix
I recently made some fixes for the destination image downloading code. the
fixes were (1) prevent crash if the Bitmap fails to return from the network and
(2) prevent a wrongthread exception. This code was used for tablet implementation
but not the phone.

This commit extracts the good downloading code and now uses it for the phone download
as well. The implementations were near mirrors anyway. just because it was hairy async
code, doesn't mean it was a good idea to duplicate implementations...