2845 tablet flights re-selection bug fix
There was a tricky bug in the code where you would select the two legs of a
roundtrip flight, and then when you go to select new flights the FlightTrip
queries were not getting reset as they should have.

The root of the problem is we were expecting this global Db.flightSearch selection
data to get reset, but the problem is this code was being housed inside of public
method of the LegsFragment. For whatever reason, those Fragments were null (and we
checked) and thus did not run the code to reset FlightSearch selection data.

In this case, we make sure to reset the data. In the event the nextLeg fragment
is present, its "resetQuery" method will take care of resetting the Db data
(and its local adapter data).