2906. Checkout Screen Total Price and Cost Summary Total Price Do Not Match on Some Flights
If the fake price change debug setting was set to "0" value, it was adding it to the price change.
Which in the case of an actual price change would not be in sync with the actual value, like as shown in Cost Summary.
Let's put in a check to make sure fakePriceChange requested is NOT zero value.