Tablet launch CPU garbage fix
Now that the launch data is dynamic, we need to account for it not being
present. This commit introduces a background color for the content pane.
It likely introduces over-draw, but we can deal with this for now as the
alternative is just plain whacky. A more complete solution will follow..