Launch screen map Otto timing fix
There is no guarantee upon the order in which Otto produce events
are fired, so we want to make sure that the LaunchCollectionsAvailable
event also stores the selected collection. This ensure that if the
CollectionClicked event is sent before the Available event, we can
make sure to draw things properly.