Improve map pin collision detection calculation
Just measure the view and ask it how big it thinks it should be. Also
forget about it completely if it overlaps, no point adding it to the
hierarchy.

Also really delay the pin popin animation until the image has loaded and
seed it with a little random start delay. Not as nice as tying it to the
position on screen but still nice.