2979 child traveler phone ui fix
The more infants than laps alert wasn't showing up properly due
to visibility of gone -> visible not redrawing. I changed it to
invisible -> visible and things worked more deterministically. This
gives us the added benefit of the guest picker ui not jumping around
as infant preferences fly in and out..