2917: Stop crashing if download resumes and data was cleared
The issue is that when ingesting a shared itinerary we go to the launch
screen with CLEAR_TOP but since the activity was already "stopped"
because it wasn't visible it doesn't rerun lifecycle to let us know it
is finishing so we don't end up cleaning up the download sitting in
BackgroundDownloader and we try to resume it once the new
`LaunchActivity` fires up. I just reworked the code to fallback instead
of continue the download. Ideally we'll be redoing the launch screen on
phone so I'm happy with a bandaid for an edgecase.