Fix tablet results rotation bug
Hotel and flight trip bucket data was being populated into Db.hotelSearch
and Db.flightSearch every time ResultsActivity.onCreate() was being invoked.
We don't want this because it is unnecessary and clears out any selection
the user has made on rotation. We make sure to populate Db.hotelSearch and
Db.flightSearch when clicking the "book now" button on the tripbucket card.