3031 - rename some parameters in SuggestionV2
The suggest API is a little whack. It is super overloaded in that
keys in the response will refer to a different concept depending
on the RegionType. As such, we had the private member "hotelId"
despite this being a "regionId" in most cases and a "hotelId"
if the RegionType was hotel. The variables have been renamed
to help clear up this confusion.