3024 flight selection bug fix
There was a problem with the FlightTripQuery wrt to the FlightFilter.
FlightTripQuery.getTrips() filters out FlightTrips by airport. The
airports in the FlightFilter are initialized with airports from
the FlightTripQuery. As such, when a FlightTripQuery is cleared
we must also clear the Filter so that it gets re-initialized with
the correct airports.