3055: Remove executePendingTransations() call
It is a very dangerous call to make too often. Tying it to the
completion of a download means we end up making the call at the wrong
time at some point and going (boom).

This breaks the flight and hotel result loading animation slightly but
I'd rather deal with that later than constant crashing now.