2950 tablet results calendar popup - clear dates
We want the popup to be interactive. If there is a date selected, show
an X that allows the user to clear that date. If you have start and end
dates selected and hit the X to clear the start date, it will clear both
the start and end date.