2950 tablet search results new popup behavior
So this is a fun one. Interaction is hammered down according to Chris. Here
is how we want shit to work:

1. Open the results screen and see the calendar, with hotels for tonight.
2. Start interacting with teh calendar (or hit hte choose flight dates button)
to show the popup. Hit done to close the popup.

Pretty simple! this required a new state CALENDAR_WITH_POPUP ... so there is a distinction
between the state where the popup blocks hotel list interaciton and the one with just
a calendar showing on load..