3026 search by hotel name - fix for phone path
We now show "Sold out on Expedia" for a given hotel row. The code
is shared between phone and tablet for HotelAdapter/TabletHotelAdapter
and HotelSummarySection. The two row layouts tend to be different, so
I wanted to update the phone row to show "Sold out on Expedia", except
it's layout behaves a little differently. I've added the special case
in here.