ResultsListFragment header TextView click cleanup
Once upon a time I added some functionality to the ResultsList
header as I was prototyping the GDE code. Now that the behavior
is hammered down and I am in this code polishing it, I figure it
best to rip out this completely unnecessary code.

Also, I took the opportunity to rename the FlightList header interface
now that it no longer is "DONE" and is instead "SORT AND FILTER".