3242: Fix showing of hotel error state
The logic was messy and hard to read. Cleaning it up showed that it was
checking the right conditions but in the wrong order so it would
fallback incorrectly to HOTEL_LIST_DOWN to show results.

Also delete all kinds of `FragmentTransaction.commit()` calls. They are
landmines waiting to explode. Especially when combined with
`executePendingTransactions()`.