Remove some disk loading code from TabletCheckoutActivity
If Sp, HotelSearch, or FlightSearch data is not in memory we won't try and
load it from disk anymore in the tablet path. We have a lot of stability
crashers that would come up from not handling the loading of this data
properly, plus the actual booking would likely fail anyway due to timeout.