3191 resume app after background kill stability fix
The app is not setup to recover nicely if Hotel/Flight/Search/TripBucket
data gets flushed from memory (like in the case of a background kill and
then resume). On the phone path we have spent energy disk caching data
and attempting to recover as best we can. On tablet we do not have
such luxury at this time.

The logic works like this: in the Results and Checkout Activity's onCreate
we check to see of the Sp.getSearchParams() has content. If the SearchParams
are empty then we have found ourselves in a background kill / resume scenario.

From here we will finish the Activity and also take extra care in their Fragments
to exit methods early if it's Activity is finishing (so as not to crash).