Tablet payment form validation fix
Suppose you try and hit done on the payment form before entering any data,
now (as of the previous commit) you will see a bunch of blue exclamation
indication marks for both the payment form and the location form. This commit
adds a section change listener to the location so that its blue marks disappear
when the user fixes them.