ISectionEditable method rename
Naming is everything. I feel like the checkout sections are so
hard to debug because the naming is a little whacky. I've renamed
the "hasValidInput" method to be a little more clear in that it is
going to perform validation on the section.

This should make it harder for a dev to want to chain validation
checks for different sections in one if block and have short-circuit
evaluation unknowingly introduce bugs.