Fixes both 3293 & 3150.
Related to sold out trip bucket item during checkout overview.
Re-iterating what George said
"If there is only one thing in the bucket (no matter if it is a hotel or a flight) we should only show the 'select another' button.
If there are two things in the bucket, show both buttons and the select again button should take you back to search results and redo the search (to eliminate the expired inventory)."