3300 - Db.getFlightSearch() behavior change
In interest of better supporting the TripBucket and disconnecting products
from search objects, we no longer use Db.getFlightSearch() for checkout
code. Db.getFlightSearch() will continue to be used for search and search
results.

Use Db.getTripBucket().getFlight() when you've arrived on a checkout screen
in order to create the itinerary, handle price changes, etc... This allows
us to better support multiple searches and bookings of products that are
not connected to the current FlightSearch sitting in Db.