3300 - flight checkout data disk persistence
There were a bunch of calls to Db.kickOffBackgroundFlightSearchSave
in the checkout code. This is no longer valid because all flight
checkout data is stored in the trip bucket. This replaces those
calls to save (and load in the event of a background kill) appropriately.

Additionally, I added some checks in the event the TripBucket failed to
load from disk properly, in which case we'll bail backwards without
crashing.