3330 - Hotel search data / TripBucket separation
This is a pretty wide-reaching change. As was recently done for flights path,
TripBucketItemHotel is now the single source of truth for hotels checkout data.
This keeps product information out of the HotelSearch class and supports a more
extensible booking model that isn't dependent upon the HotelSearch that is sitting
in Db.getHotelSearch(). This commit includes the following changes:

- Rip out "selectedAvailabilty" from HotelSearch, which was patched in to support
TripBucket bookings
- Store pertinent booking information in TripBucketItemHotel (like, CreateTripResponse,
HotelBookingResponse, Rate, etc..)
- Use Db.getTripBucket().getHotel() instead of Db.getHotelSearch() on all things
checkout (Activity, Fragment, tracking code, etc...)
- Commit a TripBucket from phone rooms/rate rate selection code