3341 - flight selection mega-sadness fix
We were persisting the "selected" FlightTrip to FlightSearch
and reading this out for all calls to getFlightTrip. This was
leftover from before the refactor and should've been refactored
along with the rest.

We no longer cache a "selected" FlightTrip in FlightSearch and instead
query for the selected flighttrip dynamically using the selected legs.
The selected FlightTrip is now stored properly inside of the TripBucketItemFlight
rather than inside of of FlightSearch. Also, this means we won't accidentally
persist an old FlightTrip and forget to clear it out and bone flight selection
until private data is cleared..