Traveler updates - goodbye areaCode (fixes 3338)
This commit removes the storage of areaCode in it's own field inside of
Phone.java. We never explicitly refer to the areaCode other than to
concatenate it with the rest of the number. Moreover, the API is beginning
to phase out returning data in areaCode (as is already done in the user/profile
information call). We just check the json at parse time and construct
the phone number accordingly.