3354: Fight the stack overflow fight
It's a really nasty crash. We end up in a loop

- setState
- register for download
- download is actually complete
- do callback immediately
- repeat

The way I solved it was I delivered the results in an Otto message so
the resulting setState happens later and we recognize we've already
processed the download.

I also cleaned up how we're doing the searches considerably. We don't
fire off Sp updates near as much as we used to and we don't trigger Sp
updates if the params are not different so the download fragments don't
need to be nearly as smart anymore.