3381 3149: Don't generate yet another full screen bitmap
SVGView is all software rendered. Worse is that it generates a
fullscreen bitmap then it is "stuck" until the activity finishes. So if
you go to results, whoops it's forever in memory.

Also this should Fix Mike's problem where his Tab 7 was trying to draw a
bitmap that was much larger than the screen and bailing out and not
showing a map at all.