3377 - can't checkout hotel after logout
When a logout occurs we now clear out hotel create trip responses. This forces
the HotelBookingFragment to perform a new create trip before booking and avoids
a TripUserNotUpdatedException.

Unfortunately, in the current architecture I believe a price change event could
occur when the user hits "FINISH BOOKING" and we kick off the create/trip call,
but we would never notify the user.