3322 don't allow selection of invalid credit cards
LCC flights don't allow booking with certain credit cards. We can't
allow the user to select an invalid credit card and attempt to book
with it. This is some feature-level stuff masked in a defect. This
commit does the following:

- Show a special icon for invalid credit cards and don't allow clicking
this row in the adapter (and don't crash for hotels..)

- Imagine the scenario where you have both a flight and hotel in your tripbucket.
Select the hotel first and select a card that happens to be invalid for the flight
that is in your TripBucket. We also clear out invalid credit cards from the BillingInfo
in this exact scenario.