3409: Goofed up billing info state
That was annoying. The bind for billing info address
was null-checking, but not "resetting" the EditTexts'
text in the case that a null list of address lines was
passed. This was goofing up the ability for the user
to recover from the scenario where they removed a stored
credit card and wanted to add a new manual CC.

Further, I added a new behavior. When a user selects a stored
credit card, it resets the working billing info. Brennan and
Akshaya were cool with this, and it removes a bunch of weird
cases related to stored CCs.