3400: Fail all credit card errors with a generic payment failure
E3 responds to some credit card failures oddly... for instance it sometimes
reports "field"="cvv" when in fact the expiration date is incorrect. This is
due to the way banks encrypt the credit card information. It's not always
possible to know which field is incorrect. Additionally, we want to give less
information to fraudsters. So, let's just be more generic about credit card
failures to solve both problems.