Re-fix Google Wallet issue
After the last fix, we checked if the client had
checked for preauthorization and if preauthorization
had already been determined to be true. If either was false,
independent of the other, we would say that Google Wallet
is still loading.

Now we check these in relation to eachother, which is what we want.

(cherry picked from commit f1daf8d80be7e0229b3bf8dd6396198c2c336016)