Re-fix Google Wallet issue
After the last fix, we checked if the client had
checked for preauthorization and if preauthorization
had already been determined to be true. If either was false,
independent of the other, we would say that Google Wallet
is still loading.

Now we check these in relation to eachother, which is what we want.