Move phone flights loading view to expedia flavor
This codebase is a little too Expedia-centric in terms of it's branding. The
gradle build system supports different product flavors pretty nicely, and as
such I've moved the phone flights loading Fragment, View, and Drawables into
the expedia sources so as not to include on other builds that don't use it.

Another benefit is that our code isn't littered with ExpediaBookingApp.IS_EXPEDIA
checks, leading to confusing code as more and more white labels are added. In this
case, each flavor provides it's own implementation of the StatusFragment. I've
encapsulated some common functionality within AbsStatusFragment.

Auditors: brennan