Check if POS locale is supported by Expedia
Summary: If both the default phone locale and POS locale are not
supported, suggest service was not returning results.  For example, if
phone locale is set to `en_ID` and POS selected as Indonesia, the
`localeIdentifier` is `en_ID` in both the cases. This does not return
any result. Defaulting to `en_US` in such cases.

Reviewers: brennan

Reviewed By: brennan

Subscribers: sugupta

Differential Revision: https://phabricator.mobiata.com/D680