3467 - Air attach spike - implement plumbing
Air attach is a feature at Expedia that offers our users discounted hotel rooms after
booking a flight. This discount window tends to last for approximately 10 days and helps
Expedia sell more hotels, probably.

The API generally takes care of keeping track of a purchased flight in the session cookies, and
from that point forward will return hotels with discounted rates. Unfortunately, the backend
does not recognize the flight itinerary instantly, and thus, we can't use the hotel API in the
standard way and take advantage of discounted air attach rooms. Therefore, we must also add
some parameters to hotel offers and create trip calls to signal the mobile API to use a cache
and potentially return the cheap rate.

The flow essentially works like this:

1. Send "useCacheForAirAttach" on hotel offers calls if we think we'll later request to use
that cache in a hotel create trip call.

2. Book a flight. The flight checkout response contains a node that signals whether or not
this session qualifies for air attach based upon the booked flight.

3. Make the hotel create trip request with "qualifyAirAttach" if the session supports air
attach after a flight booking. This create trip response will contain a discounted room
rate node that represents the updated, cheaper price.

4. Also, the "lowest rate" associated with a given property from the hotel search call contains
a flag "airAttached" that notes whether or not the lowest rate that we display on the list is
a rate that has been reduced due to air attach logic on the backend.