Attempt to fix L2ImageCache getDiskCache crash
This is a top crash from Crashlytics reporting. We suspect context.getCacheDir()
is returning null, therefore we now null check for that and fall back to another
cache directory if it is null. Additionally, the code is easier to debug in the event
we have not fixed the crash. Also, got rid of some no longer necessary compatability
code.

https://www.crashlytics.com/mobiata/android/apps/com.expedia.bookings/issues/5424b7fae3de5099baeacca6