Fix crashed related to airline names
The flight filters code for tablet computes some data dynamically. The comparator that
governs the sort order of airlines in the "airline filter" depends upon the airline name.
It looks like our FlightStatsDB does not have the airline name for every airline that we
use, so I just null-check on the airline name before blindly trying to access it.

https://www.crashlytics.com/mobiata/android/apps/com.expedia.bookings/issues/54252dc2e3de5099baeb6297