3443: Fix hotel credit card text colors
For flights and tablet checkout it just sets a solid color so using
`TextView.getCurrentTextColor()` was fine.

However for hotel credit card entry we were setting a `ColorStateList`
so we could handle focused and activated states (much like button
pressed states).

When you set a solid color on a `TextView` it encodes it into a
`ColorStateList` so it is safe to always use `TextView.getTextColors()`.