Move Route Description TextView logic to FlightController
Having this logic live in the RecursiveFlightLegsFragment relied
too heavily UI states that are ultimately controllled by the
controller. Previously, this structure would result in the
TextView's setText() being called in rapid succession, which
exposed an issue with TextViews, where multiple spans of text
overlap. These changes appear to have fixed the issue.