3537 N9 trip bucket sizing fixes
This commit stops using a hardcoded width for the trip bucket items and rather sets the width
as match_parent and uses proper margins to fill the space properly. The issue is that the N9
has a more square resolution which results in less horizontal space in landscape. There is no
need for the width to be hardcoded, so I've made this fix and the prices of trip bucket items
can be seen on the N9.