3496. Stored Travelers List Popup in checkout overview. Part1
Adding the "Saved" button for when user is logged in to help bring up the list of saved travelers.
Adding the ListPopupWindow.
On selecting any of the travelers from the list item, let's start a download to fetch the traveler details.
Added a new field "mIsSelectable" to Traveler object. This field is used to indicate if the traveler can be selected from the list or if it's grey-ed out (disabled).