Fix tablet hotels mode deep link logic
When deep linking to a specific hotel, we should like to
short-circuit the user to the hotel details view directly
on tablet results. This logic was incorrect and sending
the user to the ROOMS_RATES view even if we were doing
a generic hotel search, and only wanted to show the list
(without any property selected).