Use total price after taxes for Rate price change comparisons
The problem with using mTotalDisplayPrice is that it can be different
depending on where you get your Rate from. The API might include
mandatory fees in display price on checkout, but not rooms and rates.
As a result, any comparison between rates from these two scenarios
must account for this case. Comparing only the total with taxes,
pre-fee inclusion, is a better option for these reasons.