3762/3763 Better tablet checkout error handling
To support air attach, we secretly make a hotel create trip call in the background
before displaying the flight confirmation screen in the tablet checkout flow
if the user is booking a flight and has a hotel in the trip bucket. If the
hotel create trip call fails, we want to show the flight confirmation yet show
the trip bucket hotel item as expired. This commit does that rather than blowing
up.