3818. Saved Stored Traveler Button Disappears
Let's get rid of all the validation checkmark related code to simplify the TravelerButtonFragment.
Show the "Saved Travelers" button only when users are logged in.
Also let's reset the Db traveler before we get to the enterTraveler screen, so that when they back out TravelerButtonFragment doesnt refresh in the user's view, which might make it look like an error.