Provide AppComponent to ensure shared instances
Every time a module is constructed it results in new instances of
objects. However what we really need is to share things like the
`OkHttpClient` and `RequestInterceptor`.

This is actually really easy to see because the `AppComponent` must
declare these as publicly provided things for the `CarComponent` and
`LXComponent` to consume when being built.

One downside of this is that the `String` endpoints are now created
during app start and never updated. So we need to create an
`EndpointProvider` to hold onto the settings or the application context
and be able to spit out the current endpoint.