Fix the mysterious, top Crashlytics crasher
This one has been around since the tablet release. Hard to repro, as it
is timing related, but this will shore it up. The repro steps are to select
a flight. Leave the app and come back after it has been killed in the background.
There's a timing issue in loading of flight search data and accessing leg information.

This should be handled in a better manner, but this fix is non-invasive.

https://fabric.io/mobiata/android/apps/com.expedia.bookings/issues/550e31bd5141dcfd8f327830

(and on every release we have Crashlytics data for, not just 4.5.1)