Flight check-in notification crash
https://fabric.io/mobiata/android/apps/com.expedia.bookings/issues/550eca625141dcfd8f3386c7

One of the top remaining crashers. This is proving VERY difficult to chase down. There is fallback
code to message "Check-in for flight to Detroit" rather than "Check-in for Delta flight to Detroit",
so null-checking in this case will prevent the crash and exhibit built-in behavior.

Also, added some unit tests to help document some tricky flights-specific behavior.