Fix race condition w.r.t. phone launch list
There seems to be a case where if views are detached
from the window, and the app resumes, LaunchListWidget
doesn't register for Events prior to the posting of the
Event to start the loading animation, when necessary.
PhoneLaunchWidget has a direct reference to its list,
so just use that instead of an Event.