4691: Fix crash when logging in on flights
After logging in with an account that lets us slide to purchase right
away we end up trying to do a new create trip with a fragment that is no
longer attached to the activity.

Just ignore for now which means we won't show the point count in this
case.