Remove unnecessary persistence
Found out that we persist way too many things to disk. I believe at one
point these features were considered good however we sync this data from
the server every time we start checkout anyways so being out of sync
with the server isn't really going to happen.

It also makes it much harder to understand what is going on in classes
such as `WorkingTravelerInfoManager`. Now it is much easier to see how
and when it syncs to the server. And now we can't get corrupted from
disk in odd states either.

- TODO Cleanup `User` persistence to disk
- Remove `WorkingBillingInfo` persistence to disk
- Remove `WorkingTravelerInfo` persistence to disk
- Remove unused things (my favorite passtime these days...)