Use setText instead of typeText
From the videos and logs it looks like sometimes the way espresso injects key
events gets shut down and it can no longer type into the dialog. By using
`setText` we can work around this for now but it is not ideal, it would be
better to match the user behaviour of typing.