5433 Filter button has easily touchable click target
As far as I can tell, Toolbar assigns the action views that belong to
its menu a match_parent height equivalent to the toolbar’s minHeight,
which in this case was smaller than we wanted the filter button’s touch
area to be. By programmatically setting its touch area to match that of
the toolbar when it is shown, we ensure the user will be able to touch
it.