5650: give better error for name on card mismatches
When the entered name on the payment card does not match with
what is actually on the card, we get a distinct error code
from the server. Handle such errors differently from all other
payment card errors, indicating to the user that there is
specifically a problem with the entered name. All other payment
card errors are handled generically so as not to give too much
information to potential fraudsters.