5656: hide debit card warning for hotels on tablet
Some POS do not allow debit cards for flight purchases. In such a
case, we display a note to the user on the payment card entry form.
Unfortunately, there was no check to make sure that the item being
purchased was a flight. Added a check so that the message only
shows for flight checkout, and not on hotel checkout.