6131: prevent crash on wonky response from google geocode
In some certain scenarios, Google Geocode API call can return
unexpected results. In this case, a null is returned. The code
that handles the return value was properly handling nulls, but
kotlin was using a non-nullable type for the value, so it was
crashing instead of letting the null through.