PR Police with String Check Clauses
PR Police is a small python framework for watching changes in PRs.
It runs as part of Unit Tests PR Builder and participates in reporting.
It parses the File Patches in a PR to a structured format,
and invokes a Clause Processor, which is a registry of Clauses.
Each Clause validates a certain aspect against the File Patch,
and reports Issues which can be injected as comments in the PR.
Each Clause implements the interface IClause.
Comments injected by PR Police carry a specific signature.
PR Police erases previously injected comments in the PR on each run.