limit pr police to only small PRs
Where "small" is defined as less than 5 commits. The current logic in
PR police iterates over every commit in the PR for every file in the PR.
When the number of commits and/or files becomes large, this can easily
result in thousands of calls to the GitHub APIs which exhausts our API
allowance, breaking things like mergebot.

A more complete solution is needed here that will cache responses from the
API, but this stopgap will at least keep things from falling over while
that solution is built.