6926 - Checkout done button state for Cars and LX equivalent to Hotels.
This is to get checkout for LX and cars in parity with Hotels.
On clicking any card from the list of cards in the list takes the user directly to checkout page, these changes will ensure the user select the card and with only selecting done button user moves back to checkout page.