Topmost Flaky UI Tests
This script is in preparation for a job which will report on a daily
basis the top flaky tests which led to developers resorting to 'retest
uitests' and wasted effort. The larger goal is for teams and individuals
to take ownership and accountability for Flaky Tests. Attempting to fix these
tests will provide us an opportunity to revisit these tests and see if
these can be implemented via Robolectric itself which will be a major
win for the build pipeline and team productivity.