6835 fix credit card number being grayed out
No check was in place to determine whether the credit card number was previously grayed out on entry to the 'add credit card' view.. Therefore no code was in place to reset the text color to normal after a user left the view. This commit adds that check and resets the text color to its default color on re-entry to the view.