7318 add DOB to be displayed under completed traveler Added code to display DOB under traveler name if the traveler form is completed.
8399 show error if DOB is not valid
Added code that displays error text / error icon when a traveler has partially completed the traveler form. This change works for both flights and packages. Added image assets and error colors to resource files. Added an updateSubtitle method to handle altering the text color of the secondary text. Added string to display when traveler form is filled out incorectly, and ensured this string was consistent for single-traveler and multi-traveler trips. Added three helper methods to modularize common operations and remove redundant code in the updateStatus method.