7449. Package A11y Range sliders.
For talkback users we're changing the input method for range sliders.
Instead of one slider handling the ranges (time) let's have to seperate single sliders.
One for the upperlimit and other for lowerlimit.

Also encapsulated both input layouts into viewstub which can be inflated
depending on if talkback is enabled or not.