Fix flaky Flight Happy Path test.
The test fails only on small devices where it is not able to scroll
to legal message text and the click happens on trip total instead.
Not able to get scroll working using scrollTo in Espresso so moving
the legal messaging check before Slide to Purchase is visibile.

Scrolling works fine on small devices when we checked manually.