5715: fix PUT tracking for itin screen
We had previously only ever tracked the worst-case scenario. Now we will
track the best case scenario as well (data is already loaded and available).

Additionally, did a whole bunch of refactoring so that this feature could be
developed with TDD and to cleanup some test code.

Also fixed a long-standing bug in PUT (card 5904) that was exposed by the
new unit tests added for 5715.