4156 - Fix One Case of Transaction Too Large Exceptions
webView.saveState was saving state related to naivagation and page contents as a user
navigated through the webview.  This state would grow and grow and a user continued performing
more and more actions on the webview.  With nougat this forced a crash, transaction too large.

The whole purpose of saveState was to recover from Android killing our app.  Instead of fully recovering
we now instead just reload the initial url if the app is Killed Killed.  Crash bye bye.